requirements.txt
# API clients, store data and retrieve it
requests==2.32.5
python-dotenv==1.2.1
valyu==2.6.0
pinecone==8.1.0

# Clustering 
scikit-learn==1.8.0
numpy==2.4.2
scipy==1.17.1
hdbscan==0.8.41
umap-learn==0.5.11
